+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+✨ **New features**
+
+- New geometry engine: `GeosOpEngine` uses the `geosop` binary to perform geometry operations
+
 🐛 **Bug fixes**
 
 - `WKTParser` and `EWKTParser` now accept both `MULTIPOINT (1 2, 3 4)` and `MULTIPOINT((1 2), (3 4))` syntaxes;